requirements.txt
streamlit==1.39.0
pypdf==6.0.0
pymupdf==1.27.2
chromadb==0.5.5
sentence-transformers==3.0.1
torch==2.2.2
numpy==1.26.4
pandas==2.2.2
huggingface_hub==0.25.2
python-dotenv==1.0.1